fix tensorflow apphub example (+2 squashed commits) Squashed commits: [63f58a6] fix estimator to work with pipeline [03a6664] fix a bunch of formatting problems, bring future pipeline into 1.0dev

+++ fastestimator/backend/cross_entropy.py
@@ -1,3 +1,17 @@
+# Copyright 2019 The FastEstimator Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
 import tensorflow as tf
 import torch
 
@@ -22,5 +36,5 @@ def cross_entropy(y_pred, y_true, apply_softmax=False):
         if apply_softmax:
             ce = torch.nn.CrossEntropyLoss(reduction="none")(y_pred, y_true)
         else:
-            ce = torch.nn.NLLLoss(reduction="none")(torch.log(y_pred), y_true)
+            ce = torch.nn.NLLLoss(reduction="none")(torch.log(y_pred), y_true.long())
     return ce
